Implement automatic `$` prefixed shader props
The `u_dimensions` uniform used by both RoundedRectangleShader and DynamicShader
are automatically filled in per quad that is rendered. The was it was done originally
makes it difficult for the renderer to tell generally if two consecutive quads that
use the same shaders can be batched into the same render operation.

With `$` prefixed automatic shader props, Shaders opt-into using them and simply
set their defaults to a compatible type (for `$dimensions` it is simply a Rect).
The renderer, during addQuad() will fill our the correct values for automatic
shader props prior to rendering.

Shaders now can implement a `canBatchShaderProps()` method to compare two sets of
shader props to see if it can be batched into the same render opertion. The default
behavior is to return false. Some Shaders like the DynamicShader may opt not to
implement this at all if the overhead of comparing the two sets of shader props is too high.