Clipping
Clipping is enabled by setting the `clipping` property to true on a Node.

Clipping while a node is rotated or scaled currently leads to undesired results. More
work with transforms in general is needed before we can improve how it works with clipping.

The `clipping` example tests various clipping scenarios. A new test example utility component
PageContainer along with a special function called paginateTestRows() makes it easy to create
multi-row / multi-page test examples which we can use in the future to create regression
snapshots.

Upgrade to threadx 0.3.1 to fix issue where text does not work in ThreadX mode