Fix simple scene rendering
Simple scenes that do not involve textures, text or any later mutations were not
rendering to the canvas. I've tracked this down to a supposed double-buffer issue
involing the performance optimization that stops drawing frames when there are
no updates in the render tree detected. The Renderer was actually issusing the
appropriate WebGL vertices and draw calls for these simple scenes but it seems
a second pass at drawing is required in order for the scene to be drawn to
the canvas.

This fix simply does an extra draw/raf cycle immediately after it previously
drew updates.

Fixes #123