Fix simple scene rendering
Simple scenes that do not involve textures, text or any later mutations were not
rendering to the canvas. This is due to the default texture "ColorTexture" not
being loaded prior to the scene being rendered. This texture must be loaded
before any solid color Nodes will render. When this issue was happening
the RAF was paused right before the texture was loaded and that texture loaded
event did not trigger a requestRender() from the stage.

The solution here is to request a render once from the stage after the default
texture has loaded.