Instantiate a single clipping Rect for each Node / Text Renderer State object
The goal is to reduce the number of object allocations and hence garbage collection memory
pressure on objects that may be rapidly replaced during animations. In this case, the
clipping rect was being newly allocated each time it was changed. This change makes it
so there is only one clipping rect per Node that is persistently kept throughout the lifetime
of the Node. The SDF text renderer also keeps a persistent clipping rect for its purposes.

This does not seem to have an impact on raw FPS performance.