Feat: Allow custom data key/value storage on Nodes.
This allows the consumer of the Node's or TextNode's to inject small data properties that will be returned and/or added in the inspector `node-*` attributes.

For example:
```
  const dataNode = renderer.createNode({
    width: 505,
    height: 101,
    x: 200,
    y: 200,
    src: logo,
    parent: bg,
    data: {
      id :'dataNode',
      number: 1,
      boolean: true
    }
  });
```

In this case data: {} is an santized Map of key/values that will be stored on the main node.

*Note*: Custom data will not be passed to the CoreNode or ThreadX node. Thus data remains on the main thread and does not get forwarded to the Render thread in case of the ThreadX driver. This to avoid clashes/overhead on the ThreadX interface.

For performance reasons only strings, booleans and numbers are allowed. Passing in a nested object will not be permitted to avoid burdening Lightning nodes with excessive data.

Relates to https://github.com/lightning-js/renderer/pull/167 for `data-*` in the Inspector output.