Fix: Race condition when making stoppedResolve.
This PR moves the creation of the promise to 1 spot for both the `start()` case as well as the `loop` case, whereas `waitUntilStopped()` only returns what's created prior. Avoiding potential race conditions.

In our tests we see an animation that gets created early on during boot of the app get "stuck", never getting a promise resolve and essentially hanging the calling party.

 *The theory*:
`waitUntilStopped()` calls  `this.makeStoppedPromise();` and grabs a promise from `this.stoppedPromise`. However if `this.makeStoppedPromise();` 's creation of its `promise` is pushed to the next tick because there is a lot going on (e.g. an app is booting up) essentially not assigning `this.stoppedResolve` yet this could lead to an unresolved promise (with a short animation).

    *AND/OR*:

Both the animation `start()` AND `waitUntilStopped` are calling `this.makeStoppedPromise();` nearly simulatenously or across tick boundaries, resulting in both passing the `(this.startedResolve === null)` check and essentially overwriting each others `this.startedResolve` values leading to one overwriting the other and getting the app stuck.