VRT Dockerfile: Install the correct version of PNPM
Prior to this change, the VRT Dockerfile installed the latest
version of PNPM instead of that is specified in the project's
package.json (8.9.2 as of this writing). This led to an issue
where everytime you ran the VRT tests in --ci mode it would
run pnpm 9 which would incorrectly update the pnpm-lock.yaml
file.

This change uses `corepack` to install the project's version of
PNPM.