fix: Do not copy renderBounds to avoid boundary cross references
This fix solves repeated clipping updates where, due to the bounds being copied, the browser will optimize and use a pointer for boundaries. Causing the parent boundaries to be come intertwined with the children and the CoreNode no longer properly handling repeated clipping changes.

Also extended the viewport-events test to include automation so we can better test for it in the future.