feat: Add containBounds property to CoreNode
This commit adds the containBounds property to the CoreNode class in order to control whether child nodes are processed and rendered when a node is out of the visible area. When containBounds is enabled, out of bound nodes will not have their children processed or drawn on screen, resulting in improved performance. The default value for containBounds is true.