feat: SDF Font Face waits until CoreTextureManager is fully initialized
This fixes an issue where the SDF Font images are being uploaded while the createImageBitmap probing was still ongoing. As a result the image texture uses the lowest available createImageBitmap signature without options. Causing the alpha channel to be ignored which in turn results into jarred text rendering with SDF due to alignment issues.